requirements.txt
llama-cpp-python
lxml
mattermostdriver
requests
certifi
charset-normalizer
diskcache
idna
numpy
pip
ply
setuptools
typing-extensions
urllib3
websockets
wheel
pybars3
torch
diffusers
transformers
accelerate
gradio
